backend/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
requests==2.31.0
python-dotenv==1.0.0
earthengine-api==0.1.384
rasterio==1.3.9
geopandas==0.14.2
redis==5.0.1
cloudinary==1.38.0
joblib==1.3.2
pandas==2.1.4
httpx==0.26.0
